refactor: Remove obsolete database configurations and unnecessary files
The following obsolete database configurations and files were removed:
- MySQL: packages/create-frameup/templates/monolithic/js/mysql/db.js
- PostgreSQL: packages/create-frameup/templates/monolithic/js/postgres/db.js
- MongoDB: packages/create-frameup/templates/monolithic/js/mongodb/db.js

Utility files that are no longer in use were also removed:
- packages/create-frameup/src/utils/fileCreator.js
- packages/create-frameup/src/utils/templateLoader.js

These changes were made as part of a refactoring and cleanup of the existing code, removing redundant code and keeping only the configurations needed for the project. This improves the clarity and organization of the code, as well as its performance.

feat: The files that generate the folders according to the architecture selected by the developers were created.