feat: We create all the necessary templates or initial templates for the tools we will be providing in frameup.
We create a new function copyToolConfigs that will be in charge of copying the configurations of the selected tools to the generated project.

In the function generateScaffolding, we import and call copyToolConfigs after creating the project structure.

We update createPackageJson to include tool dependencies when the user selects certain tools.